+++ example/lib/resources/language_helper/languages/vi.dart
@@ -1,0 +1,23 @@
+final vi = {
+  ///===========================================================================
+  /// Path: ./lib/page_asset.dart
+  ///===========================================================================
+  '@path_0': './lib/page_asset.dart',
+  'Asset': 'Asset',
+  'This is the line 1': 'Đây là dòng thứ 1',
+  'This is the line 2': 'Đây là dòng thứ 2',
+
+  ///===========================================================================
+  /// Path: ./lib/main.dart
+  ///===========================================================================
+  '@path_1': './lib/main.dart',
+  'Hello': 'Xin Chào',
+  // 'Hello': 'Hello',  // Duplicated
+  // 'Hello': 'Hello',  // Duplicated
+  'Change language': 'Thay đổi ngôn ngữ',
+  'This is @number dollar': 'Đây là @number đô-la',
+  'Other Page': 'Trang Khác',
+  'Text will be changed': 'Chữ sẽ thay đổi',
+  'Text will be not changed': 'Chữ không thay đổi',
+  'This text is missing in `en`': 'Chữ này sẽ thiếu ở ngôn ngữ `en`',
+};