+++ lib/language_helper.dart
@@ -7,4 +7,5 @@ export 'src/extensions/localizations.dart';
 export 'src/extensions/serializer_extension.dart';
 export 'src/language_helper.dart';
 export 'src/models/language_conditions.dart';
+export 'src/models/language_data_provider.dart';
 export 'src/models/language_data_type.dart';